# Speed up FA Backward pass in GPU via parallelizing sequence dimension