Following the examples in https://github.com/triton-lang/triton/blob/main/python/tutorials/06-fused-attention.py and https://github.com/jax-ml/jax-triton/pull/177/files and split the dkdv and dq in backward for better seq parallel .

With this change, AxLearn Triton kernel now is much faster than before for 2k+ sequence length, and it can deliver end2end 15% speedup over old triton kernel which makes it a more palatable to use over CuDNN kernel. The exact changes in this PR are:
1. Adding seq parallel in backward pass.
2. Tune num_stage/num_warps to better suit common use case without hitting SRAM OOM.
3. Updated the benchmark test a bit to better match real prod setting.

Benchmark perf numbers: 
Set up: Batch size : 2,  Num of Heads: 16, Hidden Dim Per Head: 128
|run id |   seq_len |      XLA       |  AxLearn Old FA | Axlearn New FA |     Pallas    |  CuDNN   |
| --- | --- | --- | --- | --- | --- | --- | 
|0         |  1024.0    | 1.520580   | 1.360455       |  1.405664 | 1.334541  | 1.603672|
|1          |  2048.0   | 1.432180   | 1.827646        |  1.310446 |1.269701    |1.612260|
|2         | 4096.0    | 4.451288   | 7.259165        |0.694067   |0.639129   | 1.431808|
|3         | 8192.0     | 16.915710  | 28.574991     |  1.633464 |1.420990   | 1.052758|
|4         | 16384.0   | 61.869404|  112.639290  | 4.858109   |4.104518   | 2.245235|